requirements.txt
# Django>=5.1,<5.1.3
# djangorestframework>=3.15.0,<3.15.2
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
django-rest-knox>=5.0.0,<5.0.2
pyotp
Pillow>=8.2.0,<8.3.0
drf-spectacular>=0.15,<0.16
stripe